document keeping results from multiple experiments
If your use case of merging results is keeping values from different
experiments even if they otherwise share the same values in keyed
columns, you need to add "experiments" to the set of keys. An example
for this kind of behaviour might e.g. be section size information
gathered from binaries, which you want to keep over time.

This change documents how to achieve this.

Change-Id: Id600382e66e5a897e062250544e7b6f5f3a28949